Slow down the test in order to actually see the effect of async

+++ HeterogeneousCore/AcceleratorService/test/TestAcceleratorServiceProducerGPUHelpers.h
@@ -13,7 +13,7 @@ public:
   TestAcceleratorServiceProducerGPUTask();
   ~TestAcceleratorServiceProducerGPUTask() = default;
 
-  using ResultType = cuda::memory::device::unique_ptr<int[]>;
+  using ResultType = cuda::memory::device::unique_ptr<float[]>;
   using ResultTypeRaw = ResultType::pointer;
   
   ResultType runAlgo(int input, const ResultTypeRaw inputArray, std::function<void()> callback);
